Issue #49: Rename package of browser-engine-gecko module to mozilla.components.engine.gecko.
