+++ components/browser/engine-system/src/main/java/mozilla/components/engine/gecko/SystemEngine.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.engine.system
+package mozilla.components.engine.gecko
 
 import android.content.Context
 import android.util.AttributeSet
